+++ src/oci/database/models/autonomous_database_summary.py
@@ -618,6 +618,10 @@ class AutonomousDatabaseSummary(object):
             The value to assign to the time_maintenance_end property of this AutonomousDatabaseSummary.
         :type time_maintenance_end: datetime
 
+        :param maintenance_target_component:
+            The value to assign to the maintenance_target_component property of this AutonomousDatabaseSummary.
+        :type maintenance_target_component: str
+
         :param is_refreshable_clone:
             The value to assign to the is_refreshable_clone property of this AutonomousDatabaseSummary.
         :type is_refreshable_clone: bool
@@ -918,6 +922,7 @@ class AutonomousDatabaseSummary(object):
             'database_management_status': 'str',
             'time_maintenance_begin': 'datetime',
             'time_maintenance_end': 'datetime',
+            'maintenance_target_component': 'str',
             'is_refreshable_clone': 'bool',
             'time_of_last_refresh': 'datetime',
             'time_of_last_refresh_point': 'datetime',
@@ -1046,6 +1051,7 @@ class AutonomousDatabaseSummary(object):
             'database_management_status': 'databaseManagementStatus',
             'time_maintenance_begin': 'timeMaintenanceBegin',
             'time_maintenance_end': 'timeMaintenanceEnd',
+            'maintenance_target_component': 'maintenanceTargetComponent',
             'is_refreshable_clone': 'isRefreshableClone',
             'time_of_last_refresh': 'timeOfLastRefresh',
             'time_of_last_refresh_point': 'timeOfLastRefreshPoint',
@@ -1173,6 +1179,7 @@ class AutonomousDatabaseSummary(object):
         self._database_management_status = None
         self._time_maintenance_begin = None
         self._time_maintenance_end = None
+        self._maintenance_target_component = None
         self._is_refreshable_clone = None
         self._time_of_last_refresh = None
         self._time_of_last_refresh_point = None
@@ -3286,6 +3293,30 @@ class AutonomousDatabaseSummary(object):
         :type: datetime
         """
         self._time_maintenance_end = time_maintenance_end
+
+    @property
+    def maintenance_target_component(self):
+        """
+        Gets the maintenance_target_component of this AutonomousDatabaseSummary.
+        The component chosen for maintenance.
+
+
+        :return: The maintenance_target_component of this AutonomousDatabaseSummary.
+        :rtype: str
+        """
+        return self._maintenance_target_component
+
+    @maintenance_target_component.setter
+    def maintenance_target_component(self, maintenance_target_component):
+        """
+        Sets the maintenance_target_component of this AutonomousDatabaseSummary.
+        The component chosen for maintenance.
+
+
+        :param maintenance_target_component: The maintenance_target_component of this AutonomousDatabaseSummary.
+        :type: str
+        """
+        self._maintenance_target_component = maintenance_target_component
 
     @property
     def is_refreshable_clone(self):