+++ src/oci/database/models/autonomous_dw_database.py
@@ -608,6 +608,10 @@ class AutonomousDwDatabase(object):
             The value to assign to the time_maintenance_end property of this AutonomousDwDatabase.
         :type time_maintenance_end: datetime
 
+        :param maintenance_target_component:
+            The value to assign to the maintenance_target_component property of this AutonomousDwDatabase.
+        :type maintenance_target_component: str
+
         :param is_refreshable_clone:
             The value to assign to the is_refreshable_clone property of this AutonomousDwDatabase.
         :type is_refreshable_clone: bool
@@ -898,6 +902,7 @@ class AutonomousDwDatabase(object):
             'database_management_status': 'str',
             'time_maintenance_begin': 'datetime',
             'time_maintenance_end': 'datetime',
+            'maintenance_target_component': 'str',
             'is_refreshable_clone': 'bool',
             'time_of_last_refresh': 'datetime',
             'time_of_last_refresh_point': 'datetime',
@@ -1026,6 +1031,7 @@ class AutonomousDwDatabase(object):
             'database_management_status': 'databaseManagementStatus',
             'time_maintenance_begin': 'timeMaintenanceBegin',
             'time_maintenance_end': 'timeMaintenanceEnd',
+            'maintenance_target_component': 'maintenanceTargetComponent',
             'is_refreshable_clone': 'isRefreshableClone',
             'time_of_last_refresh': 'timeOfLastRefresh',
             'time_of_last_refresh_point': 'timeOfLastRefreshPoint',
@@ -1153,6 +1159,7 @@ class AutonomousDwDatabase(object):
         self._database_management_status = None
         self._time_maintenance_begin = None
         self._time_maintenance_end = None
+        self._maintenance_target_component = None
         self._is_refreshable_clone = None
         self._time_of_last_refresh = None
         self._time_of_last_refresh_point = None
@@ -3274,6 +3281,30 @@ class AutonomousDwDatabase(object):
         :type: datetime
         """
         self._time_maintenance_end = time_maintenance_end
+
+    @property
+    def maintenance_target_component(self):
+        """
+        Gets the maintenance_target_component of this AutonomousDwDatabase.
+        The component chosen for maintenance.
+
+
+        :return: The maintenance_target_component of this AutonomousDwDatabase.
+        :rtype: str
+        """
+        return self._maintenance_target_component
+
+    @maintenance_target_component.setter
+    def maintenance_target_component(self, maintenance_target_component):
+        """
+        Sets the maintenance_target_component of this AutonomousDwDatabase.
+        The component chosen for maintenance.
+
+
+        :param maintenance_target_component: The maintenance_target_component of this AutonomousDwDatabase.
+        :type: str
+        """
+        self._maintenance_target_component = maintenance_target_component
 
     @property
     def is_refreshable_clone(self):